refactor(domain): `Part`의 패키지 경로를 global 하위로 변경
